backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pdfminer.six==20231228
python-docx==1.1.0
python-multipart==0.0.9
